+++ src/main/kotlin/org/example/survey/dto/feedback/response/FeedbackResponse.kt
@@ -2,7 +2,7 @@ package org.example.survey.dto.feedback.response
 
 data class FeedbackResponse(
     val feedbackAt: String,
-    val formId: Int,
+    val formId: Long,
     val answers: List<AnswerResponse>,
-    val feedbackId: Int,
+    val feedbackId: Long,
 )
